fixed integer declarations

+++ contracts/ClockTowerSubscribe.sol
@@ -421,15 +421,15 @@ contract ClockTowerSubscribe is Ownable2Step {
         uint16 day;
         uint16 yearDay;
        
-        int __days = int(_days);
+        int256 __days = int(_days);
 
-        int L = __days + 68569 + 2440588;
-        int N = 4 * L / 146097;
+        int256 L = __days + 68569 + 2440588;
+        int256 N = 4 * L / 146097;
         L = L - (146097 * N + 3) / 4;
-        int _year = 4000 * (L + 1) / 1461001;
+        int256 _year = 4000 * (L + 1) / 1461001;
         L = L - 1461 * _year / 4 + 31;
-        int _month = 80 * L / 2447;
-        int _day = L - 2447 * _month / 80;
+        int256 _month = 80 * L / 2447;
+        int256 _day = L - 2447 * _month / 80;
         L = _month / 11;
         _month = _month + 2 - 12 * L;
         _year = 100 * (N - 49) + _year + L;
